datapath: configure linux node config writer through hive
Currently, the logic in charge of writing the header file containing
the local node configuration is entirely defined as part of a single
function, which has grown significantly in size, and depends on a large
set of global variables. Overall, this leads to additional friction
when extending it with new options, especially as it does not play
well with the recently introduced hive framework.

This commit implements a first step to address this issue, enabling the
possibility of providing through hive a set of extra functions returning
subsets of configurations, which are then merged to form the final
header file. Overall, this allows to break down the generation process
in smaller chunks, with each module defining its own specific settings
(e.g., based on user configurations and internal state), reducing at
the same time the dependency on global variables.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>