mwe: inject l2announcer node config through hive
This commit showcases how it is possible to inject extra node
configurations through Hive, breaking the overall logic in
smaller chunks tied to the respective modules and allowing
to reduce the usage of global variables.

Note: this does not represent a complete implementation, and
there are currently remaining blockers (AreDevicesRequired())
which prevent to fully remove the global configurations herein
mentioned.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>