maps: provide tunnel map through hive
Provide and initialize the tunnel map through hive, rather than
accessing it through the global variable. By default, the map
is created only when Cilium is configured in tunnel mode, but
additional modules can enable it if necessary.

Additionally, abstract its functionalities through an interface,
so that we can provide a fake implementation for testing purposes.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>