logging: consistent error attribute when emitted through logr
Extend the log attributes mangling logic to ensure consistency in the
attribute key used to identify an error. Indeed, the logr.Error method
takes an error as first parameter, and the slog adapter converts it into
an attribute identified as `err`. Let's replace that into `error` to
ensure consistency with all the other occurrences.

Related: d91e387bdbb6 ("cec: Switch to slog for CEC")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>