kvstore: drop obsolete removal of legacy prefixes
When running in kvstore mode, Cilium agents attempt to remove legacy
prefixes from the kvstore itself at startup. Given that this cleanup
logic has been in place for almost 7 years now [1], it is fair to
assume that it is no longer necessary at this point. Hence, let's
drop it, for simplicity.

[1]: 7ccf87701fc9 ("kvstore: New kvstore abstraction API")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>